pyblog/requirements.txt
coverage
flake8
idna==2.8
importlib-metadata==4.2.0
iniconfig==1.1.1
invoke==1.2.0
ipykernel==5.1.1
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.5.0
jmespath==0.9.4
jsonschema==3.0.1
pluggy==1.0.0
python-dotenv
py==1.11.0
pyasn1==0.4.5
pycodestyle==2.11.0
pycparser==2.19
pyflakes==2.5.0
Pygments==2.7.4
PyNaCl==1.3.0
pyparsing==3.0.9
pyrsistent==0.15.2
pytest
python-dateutil==2.8.0
pyzmq==18.0.2
requests==2.22.0
resolvelib==0.5.5
ruamel.yaml
typing-extensions==4.4.0
urllib3==1.26.5
yamllint